pyproject.toml
[project]
name = "animedatabase"
version = "0.1.0"
description = "Anime Database project for manipulation your anime titles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.8.1",
    "sqlalchemy>=2.0.40",
]

[tool.pyright]
venv = ".venv"
venvPath = "."
